Emotion-Detection-2/ResNet/dataset.py
import os
from PIL import Image
from torch.utils.data import Dataset

class AffectNetDataset(Dataset):
    def __init__(self, root_dir, transform=None):
        self.root = root_dir
        self.transform = transform
        self.samples = []
        self.class_to_idx = {}

        # Create mapping
        classes = sorted(os.listdir(root_dir))
        for idx, cls in enumerate(classes):
            self.class_to_idx[cls] = idx
            class_folder = os.path.join(root_dir, cls)

            for img in os.listdir(class_folder):
                if img.lower().endswith((".jpg", ".png", ".jpeg")):
                    self.samples.append(
                        (os.path.join(class_folder, img), idx)
                    )

    def __len__(self):
        return len(self.samples)

    def __getitem__(self, idx):
        img_path, label = self.samples[idx]
        img = Image.open(img_path).convert("RGB")

        if self.transform:
            img = self.transform(img)

        return img, label
