Emotion-Detection-2/ResNet/requirements.txt
torch
torchvision
tqdm
numpy
pillow
matplotlib
seaborn
scikit-learn
